dialogues done error in fishman
