--- Assets/Scripts/Battle.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1c9fb5888cd5bc64b85b43bde989c266
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
