--- Assets/Scripts/Battle/DamageIndicator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4b9a4256176701541acc181b35285f56
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
