--- Assets/Scripts/Battle/TurnIndicator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 71d7f66ef581772489bceb832250e2a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
