+++ Assets/Scripts/Dialogue/JoinerDialogue/Battle/BattleUiHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12da369d708793d46b121c58ee9f26e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
