+++ Assets/Scripts/Dialogue/Survivor dialogue/PTSDGuyInventoryDialogue.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 
 public class PTSDGuyInventoryDialogue : InventoryDialogues {
      public override string LowHealthAndLowHunger() {
-        return "Hey... Can we wait a bit, i think i need to catch my breath";
+        return "This is it..";
     }
     public override string LowHealthDialogue() {
-        return "Owch, do we have any potions left? i think i broke something";
+        return "I am gonna die!!!";
     }
     public override string LowHungerDialogue() {
-        return "I could use a burger dawg";
+        return "My stomach hurts like hell.";
     }
     public override string NormalDialogue() {
-        return "Woah the sun is so bright today";
+        return "Grass hasn't been this soft since I was young.";
     }
 }
